Styling enhancments - Add a line under link when page is clicked
